move signals code to own file, implemented collision
