+++ signals.py
@@ -1,0 +1,285 @@
+from simulation import Vehicle
+import time
+
+speeds = {
+    "car": 2.25,
+    "bus": 1.8,
+    "truck": 1.8,
+    "bike": 2.5,
+}  # average speeds of vehicles
+
+# Coordinates of vehicles' start
+x = {
+    "right": [0, 0, 0],
+    "down": [755, 727, 697],
+    "left": [1400, 1400, 1400],
+    "up": [602, 627, 657],
+}
+y = {
+    "right": [348, 370, 398],
+    "down": [0, 0, 0],
+    "left": [498, 466, 436],
+    "up": [800, 800, 800],
+}
+
+vehicleTypes = {0: "car", 1: "bus", 2: "truck", 3: "bike"}
+directionNumbers = {0: "right", 1: "down", 2: "left", 3: "up"}
+
+# Default values of signal timers
+defaultGreen = {0: 10, 1: 10, 2: 10, 3: 10}
+defaultRed = 150
+defaultYellow = 5
+
+signals = []
+noOfSignals = 4
+currentGreen = 0  # Indicates which signal is green currently
+nextGreen = (
+    currentGreen + 1
+) % noOfSignals  # Indicates which signal will turn green next
+currentYellow = 0  # Indicates whether yellow signal is on or off
+
+# Coordinates of signal image, timer, and vehicle count
+signalCoods = [(530, 230), (810, 230), (810, 570), (530, 570)]
+signalTimerCoods = [(530, 210), (810, 210), (810, 550), (530, 550)]
+
+# Coordinates of stop lines
+stopLines = {"right": 590, "down": 330, "left": 800, "up": 535}
+defaultStop = {"right": 580, "down": 320, "left": 810, "up": 545}
+# stops = {'right': [580,580,580], 'down': [320,320,320], 'left': [810,810,810], 'up': [545,545,545]}
+
+vehicles = {
+    "right": {0: [], 1: [], 2: [], "crossed": 0},
+    "down": {0: [], 1: [], 2: [], "crossed": 0},
+    "left": {0: [], 1: [], 2: [], "crossed": 0},
+    "up": {0: [], 1: [], 2: [], "crossed": 0},
+}
+
+# Gap between vehicles
+stoppingGap = 15  # stopping gap
+movingGap = 15  # moving gap
+
+
+class TrafficSignal:
+    def __init__(self, red, yellow, green):
+        self.red = red
+        self.yellow = yellow
+        self.green = green
+        self.signalText = ""
+
+
+class Traffic_Vehicle(Vehicle):
+    def __init__(self, lane, vehicleClass, direction_number, direction):
+        super().__init__(lane, vehicleClass, direction_number, direction)
+        vehicles[direction][lane].append(self)
+        self.index = len(vehicles[direction][lane]) - 1
+
+        if (
+            len(vehicles[direction][lane]) > 1
+            and vehicles[direction][lane][self.index - 1].crossed == 0
+        ):  # if more than 1 vehicle in the lane of vehicle before it has crossed stop line
+            if direction == "right":
+                self.stop = (
+                    vehicles[direction][lane][self.index - 1].stop
+                    - vehicles[direction][lane][self.index - 1].image.get_rect().width
+                    - stoppingGap
+                )  # setting stop coordinate as: stop coordinate of next vehicle - width of next vehicle - gap
+            elif direction == "left":
+                self.stop = (
+                    vehicles[direction][lane][self.index - 1].stop
+                    + vehicles[direction][lane][self.index - 1].image.get_rect().width
+                    + stoppingGap
+                )
+            elif direction == "down":
+                self.stop = (
+                    vehicles[direction][lane][self.index - 1].stop
+                    - vehicles[direction][lane][self.index - 1].image.get_rect().height
+                    - stoppingGap
+                )
+            elif direction == "up":
+                self.stop = (
+                    vehicles[direction][lane][self.index - 1].stop
+                    + vehicles[direction][lane][self.index - 1].image.get_rect().height
+                    + stoppingGap
+                )
+        else:
+            self.stop = defaultStop[direction]
+
+        # Set new starting and stopping coordinate
+        w_stop = self.image.get_rect().width + stoppingGap
+        h_stop = self.image.get_rect().height + stoppingGap
+        if direction == "right":
+            x[direction][lane] -= w_stop
+        elif direction == "left":
+            x[direction][lane] += w_stop
+        elif direction == "down":
+            y[direction][lane] -= h_stop
+        elif direction == "up":
+            y[direction][lane] += h_stop
+
+    def move(self):
+        if self.direction == "right":
+            if (
+                self.crossed == 0
+                and self.x + self.image.get_rect().width > stopLines[self.direction]
+            ):  # if the image has crossed stop line now
+                self.crossed = 1
+            if (
+                self.x + self.image.get_rect().width <= self.stop
+                or self.crossed == 1
+                or (currentGreen == 0 and currentYellow == 0)
+            ) and (
+                self.index == 0
+                or self.x + self.image.get_rect().width
+                < (vehicles[self.direction][self.lane][self.index - 1].x - movingGap)
+            ):
+                # (if the image has not reached its stop coordinate or has crossed stop line or has green signal) and (it is either the first vehicle in that lane or it is has enough gap to the next vehicle in that lane)
+                self.x += self.speed  # move the vehicle
+        elif self.direction == "down":
+            if (
+                self.crossed == 0
+                and self.y + self.image.get_rect().height > stopLines[self.direction]
+            ):
+                self.crossed = 1
+            if (
+                self.y + self.image.get_rect().height <= self.stop
+                or self.crossed == 1
+                or (currentGreen == 1 and currentYellow == 0)
+            ) and (
+                self.index == 0
+                or self.y + self.image.get_rect().height
+                < (vehicles[self.direction][self.lane][self.index - 1].y - movingGap)
+            ):
+                self.y += self.speed
+        elif self.direction == "left":
+            if self.crossed == 0 and self.x < stopLines[self.direction]:
+                self.crossed = 1
+            if (
+                self.x >= self.stop
+                or self.crossed == 1
+                or (currentGreen == 2 and currentYellow == 0)
+            ) and (
+                self.index == 0
+                or self.x
+                > (
+                    vehicles[self.direction][self.lane][self.index - 1].x
+                    + vehicles[self.direction][self.lane][self.index - 1]
+                    .image.get_rect()
+                    .width
+                    + movingGap
+                )
+            ):
+                self.x -= self.speed
+        elif self.direction == "up":
+            if self.crossed == 0 and self.y < stopLines[self.direction]:
+                self.crossed = 1
+            if (
+                self.y >= self.stop
+                or self.crossed == 1
+                or (currentGreen == 3 and currentYellow == 0)
+            ) and (
+                self.index == 0
+                or self.y
+                > (
+                    vehicles[self.direction][self.lane][self.index - 1].y
+                    + vehicles[self.direction][self.lane][self.index - 1]
+                    .image.get_rect()
+                    .height
+                    + movingGap
+                )
+            ):
+                self.y -= self.speed
+
+
+# Initialization of signals with default values
+def initialize():
+    ts1 = TrafficSignal(0, defaultYellow, defaultGreen[0])
+    signals.append(ts1)
+    ts2 = TrafficSignal(
+        ts1.red + ts1.yellow + ts1.green, defaultYellow, defaultGreen[1]
+    )
+    signals.append(ts2)
+    ts3 = TrafficSignal(defaultRed, defaultYellow, defaultGreen[2])
+    signals.append(ts3)
+    ts4 = TrafficSignal(defaultRed, defaultYellow, defaultGreen[3])
+    signals.append(ts4)
+    repeat()
+
+
+def repeat():
+    global currentGreen, currentYellow, nextGreen
+    while (
+        signals[currentGreen].green > 0
+    ):  # while the timer of current green signal is not zero
+        updateValues()
+        time.sleep(1)
+    currentYellow = 1  # set yellow signal on
+    # reset stop coordinates of lanes and vehicles
+    for i in range(0, 3):
+        for vehicle in vehicles[directionNumbers[currentGreen]][i]:
+            vehicle.stop = defaultStop[directionNumbers[currentGreen]]
+    while (
+        signals[currentGreen].yellow > 0
+    ):  # while the timer of current yellow signal is not zero
+        updateValues()
+        time.sleep(1)
+    currentYellow = 0  # set yellow signal off
+
+    # reset all signal times of current signal to default times
+    signals[currentGreen].green = defaultGreen[currentGreen]
+    signals[currentGreen].yellow = defaultYellow
+    signals[currentGreen].red = defaultRed
+
+    currentGreen = nextGreen  # set next signal as green signal
+    nextGreen = (currentGreen + 1) % noOfSignals  # set next green signal
+    signals[nextGreen].red = (
+        signals[currentGreen].yellow + signals[currentGreen].green
+    )  # set the red time of next to next signal as (yellow time + green time) of next signal
+    repeat()
+
+
+# Update values of the signal timers after every second
+def updateValues():
+    for i in range(0, noOfSignals):
+        if i == currentGreen:
+            if currentYellow == 0:
+                signals[i].green -= 1
+            else:
+                signals[i].yellow -= 1
+        else:
+            signals[i].red -= 1
+
+
+def step_vehicles(screen):
+    for vehicle in simulation:
+        screen.blit(vehicle.image, [vehicle.x, vehicle.y])
+        vehicle.move()
+        # remove vehicles after they drive offscreen
+
+        if not (
+            -vehicle.image.get_rect().width
+            <= vehicle.x
+            <= screen.get_width() + vehicle.image.get_rect().width
+        ):
+            vehicle.crossed = 1
+            simulation.remove(vehicle)
+            vehicles[vehicle.direction][vehicle.lane].pop(
+                vehicles[vehicle.direction][vehicle.lane] == vehicle
+            )
+        if not (
+            -vehicle.image.get_rect().height
+            <= vehicle.y
+            <= screen.get_height() + vehicle.image.get_rect().height
+        ):
+            vehicle.crossed = 1
+            simulation.remove(vehicle)
+            vehicles[vehicle.direction][vehicle.lane].pop(
+                vehicles[vehicle.direction][vehicle.lane] == vehicle
+            )
+            
+            
+class Main:
+    # thread1 = threading.Thread(
+    #     name="initialization", target=initialize, args=()
+    # )  # initialization
+    # thread1.daemon = True
+    # thread1.start()
