+++ simulation.py
@@ -4,18 +4,6 @@ import threading
 import pygame
 import sys
 
-# Default values of signal timers
-defaultGreen = {0: 10, 1: 10, 2: 10, 3: 10}
-defaultRed = 150
-defaultYellow = 5
-
-signals = []
-noOfSignals = 4
-currentGreen = 0  # Indicates which signal is green currently
-nextGreen = (
-    currentGreen + 1
-) % noOfSignals  # Indicates which signal will turn green next
-currentYellow = 0  # Indicates whether yellow signal is on or off
 
 speeds = {
     "car": 2.25,
@@ -38,38 +26,15 @@ y = {
     "up": [800, 800, 800],
 }
 
-vehicles = {
-    "right": {0: [], 1: [], 2: [], "crossed": 0},
-    "down": {0: [], 1: [], 2: [], "crossed": 0},
-    "left": {0: [], 1: [], 2: [], "crossed": 0},
-    "up": {0: [], 1: [], 2: [], "crossed": 0},
-}
 vehicleTypes = {0: "car", 1: "bus", 2: "truck", 3: "bike"}
 directionNumbers = {0: "right", 1: "down", 2: "left", 3: "up"}
 
-# Coordinates of signal image, timer, and vehicle count
-signalCoods = [(530, 230), (810, 230), (810, 570), (530, 570)]
-signalTimerCoods = [(530, 210), (810, 210), (810, 550), (530, 550)]
-
-# Coordinates of stop lines
-stopLines = {"right": 590, "down": 330, "left": 800, "up": 535}
-defaultStop = {"right": 580, "down": 320, "left": 810, "up": 545}
-# stops = {'right': [580,580,580], 'down': [320,320,320], 'left': [810,810,810], 'up': [545,545,545]}
-
 # Gap between vehicles
-stoppingGap = 15  # stopping gap
-movingGap = 15  # moving gap
+# movingGap = 15  # moving gap
+# spacings = {0: -movingGap, 1: -movingGap, 2: movingGap, 3: movingGap}
 
 pygame.init()
 simulation = pygame.sprite.Group()
-
-
-class TrafficSignal:
-    def __init__(self, red, yellow, green):
-        self.red = red
-        self.yellow = yellow
-        self.green = green
-        self.signalText = ""
 
 
 class Vehicle(pygame.sprite.Sprite):
@@ -80,261 +45,112 @@ class Vehicle(pygame.sprite.Sprite):
         self.speed = speeds[vehicleClass]
         self.direction_number = direction_number
         self.direction = direction
-        self.x = x[direction][lane]
-        self.y = y[direction][lane]
         self.crossed = 0
         path = "images/" + direction + "/" + vehicleClass + ".png"
         self.image = pygame.image.load(path)
+        self.rect = self.image.get_rect()
+        self.rect.x = x[direction][lane]
+        self.rect.y = y[direction][lane]
 
         simulation.add(self)
+
+    # def get_rect(self):
+    #     return self.image.get_rect()
 
     def render(self, screen):
         screen.blit(self.image, (self.x, self.y))
 
     def move(self):
         if self.direction == "right":
-            self.x += self.speed  # move the vehicle
+            self.rect.x += self.speed  # move the vehicle
         elif self.direction == "down":
-            self.y += self.speed
+            self.rect.y += self.speed
         elif self.direction == "left":
-            self.x -= self.speed
+            self.rect.x -= self.speed
         elif self.direction == "up":
-            self.y -= self.speed
-
-
-class Traffic_Vehicle(Vehicle):
-    def __init__(self, lane, vehicleClass, direction_number, direction):
-        super().__init__(lane, vehicleClass, direction_number, direction)
-        vehicles[direction][lane].append(self)
-        self.index = len(vehicles[direction][lane]) - 1
-
-        if (
-            len(vehicles[direction][lane]) > 1
-            and vehicles[direction][lane][self.index - 1].crossed == 0
-        ):  # if more than 1 vehicle in the lane of vehicle before it has crossed stop line
-            if direction == "right":
-                self.stop = (
-                    vehicles[direction][lane][self.index - 1].stop
-                    - vehicles[direction][lane][self.index - 1].image.get_rect().width
-                    - stoppingGap
-                )  # setting stop coordinate as: stop coordinate of next vehicle - width of next vehicle - gap
-            elif direction == "left":
-                self.stop = (
-                    vehicles[direction][lane][self.index - 1].stop
-                    + vehicles[direction][lane][self.index - 1].image.get_rect().width
-                    + stoppingGap
-                )
-            elif direction == "down":
-                self.stop = (
-                    vehicles[direction][lane][self.index - 1].stop
-                    - vehicles[direction][lane][self.index - 1].image.get_rect().height
-                    - stoppingGap
-                )
-            elif direction == "up":
-                self.stop = (
-                    vehicles[direction][lane][self.index - 1].stop
-                    + vehicles[direction][lane][self.index - 1].image.get_rect().height
-                    + stoppingGap
-                )
-        else:
-            self.stop = defaultStop[direction]
-
-        # Set new starting and stopping coordinate
-        w_stop = self.image.get_rect().width + stoppingGap
-        h_stop = self.image.get_rect().height + stoppingGap
-        if direction == "right":
-            x[direction][lane] -= w_stop
-        elif direction == "left":
-            x[direction][lane] += w_stop
-        elif direction == "down":
-            y[direction][lane] -= h_stop
-        elif direction == "up":
-            y[direction][lane] += h_stop
-
-    def move(self):
-        if self.direction == "right":
-            if (
-                self.crossed == 0
-                and self.x + self.image.get_rect().width > stopLines[self.direction]
-            ):  # if the image has crossed stop line now
-                self.crossed = 1
-            if (
-                self.x + self.image.get_rect().width <= self.stop
-                or self.crossed == 1
-                or (currentGreen == 0 and currentYellow == 0)
-            ) and (
-                self.index == 0
-                or self.x + self.image.get_rect().width
-                < (vehicles[self.direction][self.lane][self.index - 1].x - movingGap)
-            ):
-                # (if the image has not reached its stop coordinate or has crossed stop line or has green signal) and (it is either the first vehicle in that lane or it is has enough gap to the next vehicle in that lane)
-                self.x += self.speed  # move the vehicle
-        elif self.direction == "down":
-            if (
-                self.crossed == 0
-                and self.y + self.image.get_rect().height > stopLines[self.direction]
-            ):
-                self.crossed = 1
-            if (
-                self.y + self.image.get_rect().height <= self.stop
-                or self.crossed == 1
-                or (currentGreen == 1 and currentYellow == 0)
-            ) and (
-                self.index == 0
-                or self.y + self.image.get_rect().height
-                < (vehicles[self.direction][self.lane][self.index - 1].y - movingGap)
-            ):
-                self.y += self.speed
-        elif self.direction == "left":
-            if self.crossed == 0 and self.x < stopLines[self.direction]:
-                self.crossed = 1
-            if (
-                self.x >= self.stop
-                or self.crossed == 1
-                or (currentGreen == 2 and currentYellow == 0)
-            ) and (
-                self.index == 0
-                or self.x
-                > (
-                    vehicles[self.direction][self.lane][self.index - 1].x
-                    + vehicles[self.direction][self.lane][self.index - 1]
-                    .image.get_rect()
-                    .width
-                    + movingGap
-                )
-            ):
-                self.x -= self.speed
-        elif self.direction == "up":
-            if self.crossed == 0 and self.y < stopLines[self.direction]:
-                self.crossed = 1
-            if (
-                self.y >= self.stop
-                or self.crossed == 1
-                or (currentGreen == 3 and currentYellow == 0)
-            ) and (
-                self.index == 0
-                or self.y
-                > (
-                    vehicles[self.direction][self.lane][self.index - 1].y
-                    + vehicles[self.direction][self.lane][self.index - 1]
-                    .image.get_rect()
-                    .height
-                    + movingGap
-                )
-            ):
-                self.y -= self.speed
+            self.rect.y -= self.speed
 
 
 class RL_Vehicle(Vehicle):
     def __init__(self, lane, vehicleClass, direction_number, direction):
         super().__init__(lane, vehicleClass, direction_number, direction)
-        self.crossed = 1
-
-
-# Initialization of signals with default values
-def initialize():
-    ts1 = TrafficSignal(0, defaultYellow, defaultGreen[0])
-    signals.append(ts1)
-    ts2 = TrafficSignal(
-        ts1.red + ts1.yellow + ts1.green, defaultYellow, defaultGreen[1]
-    )
-    signals.append(ts2)
-    ts3 = TrafficSignal(defaultRed, defaultYellow, defaultGreen[2])
-    signals.append(ts3)
-    ts4 = TrafficSignal(defaultRed, defaultYellow, defaultGreen[3])
-    signals.append(ts4)
-    repeat()
-
-
-def repeat():
-    global currentGreen, currentYellow, nextGreen
-    while (
-        signals[currentGreen].green > 0
-    ):  # while the timer of current green signal is not zero
-        updateValues()
-        time.sleep(1)
-    currentYellow = 1  # set yellow signal on
-    # reset stop coordinates of lanes and vehicles
-    for i in range(0, 3):
-        for vehicle in vehicles[directionNumbers[currentGreen]][i]:
-            vehicle.stop = defaultStop[directionNumbers[currentGreen]]
-    while (
-        signals[currentGreen].yellow > 0
-    ):  # while the timer of current yellow signal is not zero
-        updateValues()
-        time.sleep(1)
-    currentYellow = 0  # set yellow signal off
-
-    # reset all signal times of current signal to default times
-    signals[currentGreen].green = defaultGreen[currentGreen]
-    signals[currentGreen].yellow = defaultYellow
-    signals[currentGreen].red = defaultRed
-
-    currentGreen = nextGreen  # set next signal as green signal
-    nextGreen = (currentGreen + 1) % noOfSignals  # set next green signal
-    signals[nextGreen].red = (
-        signals[currentGreen].yellow + signals[currentGreen].green
-    )  # set the red time of next to next signal as (yellow time + green time) of next signal
-    repeat()
-
-
-# Update values of the signal timers after every second
-def updateValues():
-    for i in range(0, noOfSignals):
-        if i == currentGreen:
-            if currentYellow == 0:
-                signals[i].green -= 1
-            else:
-                signals[i].yellow -= 1
-        else:
-            signals[i].red -= 1
+        # self.crossed = 1
 
 
 # Generating vehicles in the simulation
 def generateVehicles():
     while True:
         vehicle_type = random.randint(0, 3)
-        lane_number = random.randint(1, 2)
+        # lane_number = random.randint(1, 2)
+
+        lane_number = 1
         temp = random.randint(0, 99)
         direction_number = 0
-        dist = [25, 50, 75, 100]
-        if temp < dist[0]:
+        # dist = [25, 50, 75, 100]
+        # if temp < dist[0]:
+        #     direction_number = 0
+        # elif temp < dist[1]:
+        #     direction_number = 1
+        # elif temp < dist[2]:
+        #     direction_number = 2
+        # elif temp < dist[3]:
+        #     direction_number = 3
+        dist = 50
+        if temp < 50:
             direction_number = 0
-        elif temp < dist[1]:
-            direction_number = 1
-        elif temp < dist[2]:
+        else:
             direction_number = 2
-        elif temp < dist[3]:
-            direction_number = 3
         Vehicle(
             lane_number,
             vehicleTypes[vehicle_type],
             direction_number,
             directionNumbers[direction_number],
         )
-        time.sleep(0.5)
+        time.sleep(2)
+
+
+def step_vehicles(screen):
+    for vehicle in simulation:
+        old_x = vehicle.rect.x
+        old_y = vehicle.rect.y
+        vehicle.move()
+        # remove vehicles after they drive offscreen
+        if not (
+            -vehicle.rect.width
+            <= vehicle.rect.x
+            <= screen.get_width() + vehicle.rect.width
+        ):
+            simulation.remove(vehicle)
+        if not (
+            -vehicle.rect.height
+            <= vehicle.rect.y
+            <= screen.get_height() + vehicle.rect.height
+        ):
+            simulation.remove(vehicle)
+        # check if vehicles collided
+        collided = pygame.sprite.spritecollide(vehicle, simulation, False)
+        if len(collided) > 1:  # if they are move them back
+
+            vehicle.rect.x = old_x
+            vehicle.rect.y = old_y
+            # or destroy them if collision at spawn
+            if (
+                old_x == 0.0
+                or old_y == 0.0
+                or old_x == screen.get_width()
+                or old_y == screen.get_height()
+            ):
+                pygame.sprite.spritecollide(vehicle, simulation, True)
+
+        screen.blit(vehicle.image, [vehicle.rect.x, vehicle.rect.y])
 
 
 class Main:
-    # thread1 = threading.Thread(
-    #     name="initialization", target=initialize, args=()
-    # )  # initialization
-    # thread1.daemon = True
-    # thread1.start()
-
-    # Colours
-    black = (0, 0, 0)
-    white = (255, 255, 255)
-
-    # Screensize
-    screenWidth = 1400
-    screenHeight = 800
-    screenSize = (screenWidth, screenHeight)
 
     # Setting background image i.e. image of intersection
     background = pygame.image.load("images/intersection.png")
 
-    screen = pygame.display.set_mode(screenSize)
+    screen = pygame.display.set_mode((1400, 800))
     pygame.display.set_caption("SIMULATION")
 
     # Loading signal images and font
@@ -344,13 +160,18 @@ class Main:
     font = pygame.font.Font(None, 30)
 
     # add RL agent
-    simulation.add(RL_Vehicle(1, "bike", 3, "up"))
+    # simulation.add(RL_Vehicle(1, "bike", 3, "up"))
 
+    # background thread to generate traffic
     thread2 = threading.Thread(
         name="generateVehicles", target=generateVehicles, args=()
-    )  # Generating vehicles
+    )
     thread2.daemon = True
     thread2.start()
+
+    # init vehicle sprites
+    for vehicle in simulation:
+        screen.blit(vehicle.image, [vehicle.rect.x, vehicle.rect.y])
 
     while True:
         for event in pygame.event.get():
@@ -358,57 +179,10 @@ class Main:
                 sys.exit()
 
         screen.blit(background, (0, 0))  # display background in simulation
-        # for i in range(
-        #     0, noOfSignals
-        # ):  # display signal and set timer according to current status: green, yello, or red
-        #     if i == currentGreen:
-        #         if currentYellow == 1:
-        #             signals[i].signalText = signals[i].yellow
-        #             screen.blit(yellowSignal, signalCoods[i])
-        #         else:
-        #             signals[i].signalText = signals[i].green
-        #             screen.blit(greenSignal, signalCoods[i])
-        #     else:
-        #         if signals[i].red <= 10:
-        #             signals[i].signalText = signals[i].red
-        #         else:
-        #             signals[i].signalText = "---"
-        #         screen.blit(redSignal, signalCoods[i])
-        signalTexts = ["", "", "", ""]
 
-        # display signal timer
-        # for i in range(0, noOfSignals):
-        #     signalTexts[i] = font.render(str(signals[i].signalText), True, white, black)
-        #     screen.blit(signalTexts[i], signalTimerCoods[i])
-
+        step_vehicles(screen)
         # display the vehicles
-        for vehicle in simulation:
-            screen.blit(vehicle.image, [vehicle.x, vehicle.y])
-            vehicle.move()
-            # remove vehicles after they drive offscreen
 
-            # if not (
-            #     -vehicle.image.get_rect().width
-            #     <= vehicle.x
-            #     <= screenWidth + vehicle.image.get_rect().width
-            # ):
-            #     print([vehicle.x, vehicle.y])
-            #     vehicle.crossed = 1
-            #     simulation.remove(vehicle)
-            #     vehicles[vehicle.direction][vehicle.lane].pop(
-            #         vehicles[vehicle.direction][vehicle.lane] == vehicle
-            #     )
-            # if not (
-            #     -vehicle.image.get_rect().height
-            #     <= vehicle.y
-            #     <= screenHeight + vehicle.image.get_rect().height
-            # ):
-            #     print([vehicle.x, vehicle.y])
-            #     vehicle.crossed = 1
-            #     simulation.remove(vehicle)
-            #     vehicles[vehicle.direction][vehicle.lane].pop(
-            #         vehicles[vehicle.direction][vehicle.lane] == vehicle
-            #     )
         pygame.display.update()
